+++ system/modules/listing/modules/ModuleListing.php
@@ -192,12 +192,14 @@ class ModuleListing extends \Module
 		$strQuery .= $strWhere;
 
 		// Cast date fields to int (see #5609)
-		$blnCastInt = ($GLOBALS['TL_DCA'][$this->list_table]['fields'][$this->list_sort]['eval']['rgxp'] == 'date' || $GLOBALS['TL_DCA'][$this->list_table]['fields'][$this->list_sort]['eval']['rgxp'] == 'time' || $GLOBALS['TL_DCA'][$this->list_table]['fields'][$this->list_sort]['eval']['rgxp'] == 'datim');
+		$isInt = function($field) {
+			return $GLOBALS['TL_DCA'][$this->list_table]['fields'][$field]['eval']['rgxp'] == 'date' || $GLOBALS['TL_DCA'][$this->list_table]['fields'][$field]['eval']['rgxp'] == 'time' || $GLOBALS['TL_DCA'][$this->list_table]['fields'][$field]['eval']['rgxp'] == 'datim';
+		};
 
 		// Order by
 		if (\Input::get('order_by'))
 		{
-			if ($blnCastInt)
+			if ($isInt(\Input::get('order_by')))
 			{
 				$strQuery .= " ORDER BY CAST(" . \Input::get('order_by') . " AS SIGNED) " . \Input::get('sort');
 			}
@@ -208,7 +210,7 @@ class ModuleListing extends \Module
 		}
 		elseif ($this->list_sort)
 		{
-			if ($blnCastInt)
+			if ($isInt($this->list_sort))
 			{
 				$strQuery .= " ORDER BY CAST(" . $this->list_sort . " AS SIGNED)";
 			}