+++ system/modules/core/library/Contao/Combiner.php
@@ -328,13 +328,24 @@ class Combiner extends \System
 		if ($arrFile['extension'] == self::SCSS)
 		{
 			$objCompiler = new \scssc();
-			$objCompiler->setImportPaths(TL_ROOT . '/' . dirname($arrFile['name']));
+
+			$objCompiler->setImportPaths(array
+			(
+				TL_ROOT . '/' . dirname($arrFile['name']),
+				TL_ROOT . '/assets/compass/' . $GLOBALS['TL_ASSETS']['COMPASS']
+			));
+
 			$objCompiler->setFormatter((\Config::get('debugMode') ? 'scss_formatter' : 'scss_formatter_compressed'));
 		}
 		else
 		{
 			$objCompiler = new \lessc();
-			$objCompiler->setImportDir(TL_ROOT . '/' . dirname($arrFile['name']));
+
+			$objCompiler->setImportDir(array
+			(
+				TL_ROOT . '/' . dirname($arrFile['name'])
+			));
+
 			$objCompiler->setFormatter((\Config::get('debugMode') ? 'lessjs' : 'compressed'));
 		}